requirements.txt
dnspython==2.2.1
pandas==1.4.2
requests==2.28.0
